IdentityService Added

+++ src/AuctionService/AuctionService/Controllers/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -43,6 +44,7 @@ namespace AuctionService.Controllers
             }
             return _mapper.Map<AuctionDto>(auction);
         }
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto createAuctionDto)
         {
@@ -57,7 +59,7 @@ namespace AuctionService.Controllers
             auction.AuctionEnd = createAuctionDto.AuctionEnd;
             auction.ReservePrice = createAuctionDto.ReservePrice;
             auction.Status = Status.Live;
-            auction.Seller = "test";
+            auction.Seller = User.Identity.Name;
 
             _context.Auctions.Add(auction);
 
@@ -73,12 +75,15 @@ namespace AuctionService.Controllers
             
             return CreatedAtAction(nameof(GetAuction), new { id = auction.Id },newAuction);
         }
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
         {
             var auction = await _context.Auctions.Include(x => x.Item).FirstOrDefaultAsync(x=>x.Id == id);
             
             if (auction == null)  return NotFound();
+
+            if (auction.Seller != User.Identity.Name) return Forbid();
             
             auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
             auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -86,18 +91,26 @@ namespace AuctionService.Controllers
             auction.Item.Year = updateAuctionDto.Year ?? auction.Item.Year;
             auction.Item.Mileage = updateAuctionDto.Mileage ?? auction.Item.Mileage;
 
+            await _publishEndpoint.Publish(_mapper.Map<AuctionUpdated>(auction));
+
             var result = await _context.SaveChangesAsync() > 0;
 
             if (!result) return BadRequest("Cloud not save changes to the DB");
             
             return Ok();
         }
+
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteAuction(Guid id)
         {
             var auction = await _context.Auctions.FindAsync(id);
             if (auction == null) return NotFound();
+
+            if (auction.Seller != User.Identity.Name) return Forbid();
+
             _context.Auctions.Remove(auction);
+            await _publishEndpoint.Publish<AuctionDeleted>(new AuctionDeleted { Id = auction.Id.ToString()});
             var result = await _context.SaveChangesAsync() > 0;
             if (!result) return BadRequest("Cloud not delete to item");
             return Ok();
